Pylint tested on this file
